// src/types/PrintShop.ts
export interface Printshop {
    id: string;
    idInt: number;
    proprietario: null;
    pixGateway: null;
    cartaoGateway: null;
    idSellerZoop: null;
    statusZoop: null;
    idSellerAsaas: null;
    keySellerAsaas: null;
    statusAsaas: null;
    emailGateway: null;
    contaGateway: null;
    documentos: null;
    idCompanyDesk: null;
    tipo: string;
    ramo: string;
    outro: null;
    tipoRegEmpresa: null;
    razao: string;
    fantasia: string;
    cnpj: string;
    ie: string;
    im: string;
    dataAbertura: null;
    nome: null;
    sobrenome: null;
    sexo: null;
    cpf: null;
    rg: null;
    dataNascimento: null;
    email: null;
    site: null;
    fone1: string;
    fone2: null;
    facebook: null;
    twitter: null;
    descricao: null;
    nomeFatura: null;
    mcc: null;
    observacoes: null;
    ativo: boolean;
    bloqueado: boolean;
    enderecos: any[];
    contatos: null;
    modulosGrafica: null;
    plano: null;
    precoPlano: null;
    qtdUsuarios: null;
    qtdChat: null;
    qtdNfe: null;
    qtdNfce: null;
    qtdNfse: null;
    qtdCte: null;
    qtdFiles: null;
    qtdCtrs: null;
    qtdCnpjs: number;
    logo: string;
    rectangularLogo: null;
    faturamento: null;
    usuarioPrincipal: null;
    excluido: null;
    certificado: null;
    tipoEmpresa: null;
    nomeCidadeTecnospeed: null;
    isento: boolean;
    cedenteId: null;
    status: string;
    dataCadastro: null;
    dataPagante: null;
    dataStatus: null;
    locale: string;
    dateUpdateLogNf: null;
    syncUpdateLogNf: null;
    mei: null;
    products: null;
    generatingParcels: boolean;
  }